Added magnitude calculation for planets
BugFIX: Object coordinates were used instead of solar coordinates for the calculation of phase angle of a celestial object.

Added the calculation of the magnitude of the planets with the exception of Saturn.

Moved pc and AU to a new Units struct that should contain all unit conversions for now.

Moved the reciprocal of rpi as used to convert radians into degrees to the Units struct. Should refactor and remove all mentions of Double.rpi